Most of the class components could be written as functional
