--- src/css/cityForecastHeader.css
@@ -1,15 +1,0 @@
-.header {
-    margin: 25px 30px;
-    font-size: 28px;
-    display: inline-block;
-}
-
-.header .button-back {
-    color: rgb(78, 102, 112);
-    margin-right: 15px;
-    vertical-align: middle;
-    padding: 0;
-    background: none;
-    outline: none;
-    border: none;
-}
