--- src/css/currentWeather.css
@@ -1,78 +1,0 @@
-.current-weather {
-    margin: 0 10px 0 20px;
-}
-
-.current-weather h3 {
-    font-size: 24px;
-    font-weight: 300;
-    padding: 0;
-    margin: 0;
-}
-
-.current-weather h4 {
-    margin: 10px 0;
-    padding: 0;
-    font-size: 20px;
-    font-weight: 300;
-}
-
-.current-weather h1 {
-    color: rgb(227, 150, 129);
-    display: block;
-    text-align: center;
-    font-size: 80px;
-    line-height: 150px;
-    margin: 0;
-    padding: 0;
-    font-weight: normal;
-}
-
-.current-weather .weather-icon {
-    margin: 0 25px;
-}
-
-.current-weather ul {
-    color: rgb(227, 150, 129);
-    list-style: none;
-    padding: 0;
-    margin: 0;
-}
-
-.current-weather .weather-sliced li {
-    margin: 5px 4%;
-    font-size: 18px;
-    width: 40%;
-    display: inline-block;
-}
-
-.current-weather .weather-sliced li > * {
-    display: inline-block;
-}
-
-.current-weather .weather-sliced li > .slice-caption {
-    width: 80px;
-}
-
-.wi-day-clear:before, .wi-clear:before {
-    content:"\f00d"
-}
-
-@media screen and (min-width: 568px) {
-    .current-weather .weather-description {
-        display: inline-block !important;
-        text-align: left !important;
-    }
-
-    .current-weather .weather-sliced {
-        margin: 0 0 0 20px !important;
-        display: inline-block !important;
-        max-width: 150px !important;
-    }
-
-    .current-weather .weather-sliced li {
-        font-size: 16px !important;
-        width: auto !important;
-        margin: 10px 0 !important;
-        display: block !important;
-    }
-}
